+++ plugins/pagerduty/src/components/PagerDutyCard/index.test.tsx
@@ -15,7 +15,7 @@
  */
 import React from 'react';
 import { render, waitFor, fireEvent, act } from '@testing-library/react';
-import { PagerDutyCard } from './PagerDutyCard';
+import { PagerDutyCard } from '../PagerDutyCard';
 import { Entity } from '@backstage/catalog-model';
 import { EntityProvider } from '@backstage/plugin-catalog-react';
 import { wrapInTestApp } from '@backstage/test-utils';
@@ -25,8 +25,8 @@ import {
   ApiRegistry,
   createApiRef,
 } from '@backstage/core';
-import { pagerDutyApiRef, UnauthorizedError, PagerDutyClient } from '../api';
-import { Service } from './types';
+import { pagerDutyApiRef, UnauthorizedError, PagerDutyClient } from '../../api';
+import { Service } from '../types';
 
 const mockPagerDutyApi: Partial<PagerDutyClient> = {
   getServiceByIntegrationKey: async () => [],
